fix: address bugbot concerns

+++ libs/shared/shared/owner_data_export/config.py
@@ -12,46 +12,63 @@ EXPORT_DAYS_DEFAULT = 60
 BATCH_SIZE = 1000
 
 # Default 7 days
-DOWNLOAD_URL_EXPIRY_SECONDS = get_config(
-    "services", "owner_export", "download_url_expiry_seconds", default=7 * 24 * 60 * 60
+DOWNLOAD_URL_EXPIRY_SECONDS = int(
+    get_config(
+        "services",
+        "owner_export",
+        "download_url_expiry_seconds",
+        default=7 * 24 * 60 * 60,
+    )
 )
 
-SQL_TASK_SOFT_TIME_LIMIT = get_config(
-    "services",
-    "owner_export",
-    "sql_task_soft_time_limit",
-    default=3600,  # 1 hour
+SQL_TASK_SOFT_TIME_LIMIT = int(
+    get_config(
+        "services",
+        "owner_export",
+        "sql_task_soft_time_limit",
+        default=3600,  # 1 hour
+    )
 )
-SQL_TASK_TIME_LIMIT = get_config(
-    "services",
-    "owner_export",
-    "sql_task_time_limit",
-    default=3660,  # 1 hour + 1 minute buffer
+SQL_TASK_TIME_LIMIT = int(
+    get_config(
+        "services",
+        "owner_export",
+        "sql_task_time_limit",
+        default=3660,
+    )
 )
-ARCHIVE_TASK_SOFT_TIME_LIMIT = get_config(
-    "services",
-    "owner_export",
-    "archive_task_soft_time_limit",
-    default=7200,  # 2 hours
+ARCHIVE_TASK_SOFT_TIME_LIMIT = int(
+    get_config(
+        "services",
+        "owner_export",
+        "archive_task_soft_time_limit",
+        default=7200,  # 2 hours
+    )
 )
-ARCHIVE_TASK_TIME_LIMIT = get_config(
-    "services",
-    "owner_export",
-    "archive_task_time_limit",
-    default=7260,  # 2 hours + 1 minute buffer
+ARCHIVE_TASK_TIME_LIMIT = int(
+    get_config(
+        "services",
+        "owner_export",
+        "archive_task_time_limit",
+        default=7260,
+    )
 )
 
-FINALIZE_TASK_SOFT_TIME_LIMIT = get_config(
-    "services",
-    "owner_export",
-    "finalize_task_soft_time_limit",
-    default=3600,  # 1 hour
+FINALIZE_TASK_SOFT_TIME_LIMIT = int(
+    get_config(
+        "services",
+        "owner_export",
+        "finalize_task_soft_time_limit",
+        default=3600,  # 1 hour
+    )
 )
-FINALIZE_TASK_TIME_LIMIT = get_config(
-    "services",
-    "owner_export",
-    "finalize_task_time_limit",
-    default=3660,  # 1 hour + 1 minute buffer
+FINALIZE_TASK_TIME_LIMIT = int(
+    get_config(
+        "services",
+        "owner_export",
+        "finalize_task_time_limit",
+        default=3660,
+    )
 )
 
 # Number of concurrent workers for archive file copying
